Patin_Clement_1_application_052024/requirements.txt
# azure_storage==0.37.0
Flask==3.0.3
joblib==1.4.2
numpy==1.26.4
pandas==2.2.2
Requests==2.32.2
scikit_surprise==1.1.3
scikit_learn==1.4.2